Changed 'double' into 'decimal' in most places.

+++ client/AssetsCache/Models/AssetPair.cs
@@ -9,8 +9,8 @@ namespace AssetsCache
         public int InvertedAccuracy { get; set; }
         public string BaseAssetId { get; set; }
         public string QuotingAssetId { get; set; }
-        public double MinVolume { get; set; }
-        public double MinInvertedVolume { get; set; }
+        public decimal MinVolume { get; set; }
+        public decimal MinInvertedVolume { get; set; }
         public bool IsDisabled { get; set; }
     }
 }
